+++ include/mesh/elements/elements.hpp
@@ -1,0 +1,7 @@
+#ifndef _ELEMENTS_HPP
+#define _ELEMENTS_HPP
+
+#include "axial_elements.hpp"
+#include "tangential_elements.hpp"
+
+#endif
